[HOTFIX puyo_env] fix trimming before listing
The previous codes changes NN values whose first procedure matches the second procedure of the previous chosen procedure transition.
The size of the NN values would be replaced is eqaul to num_candidate with single depth, which basically over the num_candidate_threshold.
This means that sometimes the previously chosen transition is trimmed because it's replaced NN values is as same as the other replaced NN values.

To avoid this case, the current codes only replace the NN value with full-match transition.